refactor: move slashdo command loading out of fileUtils into its own lib module
fileUtils.js is a generic file-utilities module (JSON/JSONL IO, path-safe
resolvers, MIME detection, upload staging), but it also carried slashdo
command/lib loading and template resolution (include inlining, if:teams
conditional resolution, resolved-body staging) — one cohesive feature, not
a generic file helper. Moved loadSlashdoFile, loadSlashdoLib,
writeResolvedSlashdoBody, and their private helpers (slashdoIncludeName,
resolveSlashdoIncludes, resolveSlashdoConditionals) verbatim into a new
server/lib/slashdoLoader.js.

Chose a new module over folding into the existing slashdoInvocation.js
(which resolves how a workflow is TYPED per host CLI) because the moved
code is sizable (~200 lines) and a distinct concern — loading/resolving a
body vs. resolving its invocation shape. slashdoLoader.js imports
atomicWrite/PATHS back from fileUtils.js; no import cycle exists since
slashdoInvocation.js only imports providerModels.js and tuiHandshake.js.

No backward-compatible re-export was left in fileUtils.js: server/lib/
index.js is a flat `export *` barrel, and its collision-detector test
fails the build if two flat-exported modules in the same directory export
the same identifier. All consumers (subAgentSpawner.js, taskPromptService.js,
agentPromptBuilder.js, and their tests) were updated to import from the new
module directly instead.

(cherry picked from commit 7e4a7d45855f1348b7c25ac0f3235201d5ea6f4a)
