+++ server/services/subAgentSpawner.js
@@ -17,7 +17,8 @@ import { emitLog, cosEvents } from './cosEvents.js';
 import { updateAgent, completeAgent } from './cosAgents.js';
 import { initProviderStatus } from './providerStatus.js';
 import { onCosRunnerEvent, initCosRunnerConnection, isRunnerAvailable } from './cosRunnerClient.js';
-import { loadSlashdoFile, PATHS } from '../lib/fileUtils.js';
+import { PATHS } from '../lib/fileUtils.js';
+import { loadSlashdoFile } from '../lib/slashdoLoader.js';
 import { getRunnerOutputBatcher, flushRunnerOutputBatcher } from './agentRunnerOutputBatchers.js';
 import { syncRunnerAgents } from './agentRunnerSync.js';
 import { handleAgentCompletion, spawnAgentForTask } from './agentLifecycle.js';
